pyproject.toml
# pyproject.toml

[build-system]
requires = ["setuptools>=64.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "userstory8020"
version = "0.0.35"
description = "test assignment"
readme = "README.md"
authors = [{ name = "Mykola Kis", email = "sikalokym@gmail.com" }]

dependencies = [
    "pyspark == 3.4.1",
    "pandas == 1.3.5",
    "pyarrow == 12.0.1",
    "grpcio == 1.57.0",
    "google == 3.0.0",
    "google-api-core == 2.11.1",
    "grpcio-status == 1.57.0",
    "PyYAML == 6.0.1"
]

[project.scripts]
userstory8020 = "userstory8020.__main__:main"

[tool.setuptools.packages.find]
include = ["userstory8020"]



